Added protection for cross-site request forgery.
Consists of a hidden field containing a form key that is checked on submitting.
The key has a timeout of 6 hours (meaning the form must be submitted within 6 hours after the key generation) or the end of session if inferior, and is built using the current server host, magic string, and time.

Drawback: the session must have been started for a key enabled form to be displayed or checked.
The key is enabled by default, but can be disabled by setting <formkey>0</formkey> in the <form> element.