Added the ability to define parameters to be added to links in templates in a semi-automatic way.
Parameters are first added to the template, then, all the links
created within the template using the mkLink method will have
these parameters added.

It is a shortcut to the setting of these variables and their display
again and again in all links in the template. It also make the
template much more readable when you need lots of link with the
same parameters.