New weePDFOutput driver
It allows the easy and integrated PDF generation
using 'pdflatex' available from any LaTeX distribution. LaTeX exists on
all popular and less popular platforms.

Using this tool allow us to create PDF files by writing LaTeX templates.
The main advantage of this is that you can use the templating system
already present in wee to create a PDF file.

In order to tell wee that you want to create a PDF, you must call
weePDFOutput::select() before the end of your event in the frame.
This will select the PDF driver, which will automatically convert
your template (that you will write in LaTeX) to PDF and send it to
the browser.

In order to make this work a few changes have been made to the output
drivers. The output driver is no longer a singleton, it is rather now
a driver that you can change at will during the execution. This allow
us to define a default driver, XHTML, and use the PDF driver only for
a small subset of the frames, by selecting it as explained above.

Another change is that the code to determine the cookies path is no
longer called each time, but rather when we need to set cookies, saving
time for when we don't need it.