Added email templates.
Allows the use of template files to define email body and common headers directly
in the template file instead of doing that in the PHP file. Using the templating
system also add the benefit of reusing easily the files, using sub-templates,
and better security by stripping automatically the NUL characters using the
weeTextOutput output. The templates currently only allow for text output.

weeTextOutput is a new output method for plain text output. It is currently
only used for emails but can be selected for other uses like with any other
output method by calling select().

The template file has two parts: first the headers, then an empty line to
separate the parts, and then the body. If there is no header defined, you MUST
begin the template file with the empty line. It should not happen often since you
SHOULD define a Subject header anyway.