New example application: Pastebin
Will be used as an example to write the chapter "Starting a New Application" of the documentation.
Features 3 frames: the `toppage` for some explanations and posting a new Pastebin,
the `view` to access a specific Pastebin, and the `last` displays the 5 last posted.
`toppage` demonstrates how to create and use forms, `view` how to query and fetch
exactly one row, and `last` how to query and fetch more than one row.