weeFileConfig include instructions
weeFileConfig now supports include instructions, you just need to write ".include another_config_file.cnf" to include it.
The include mechanism stores the stack of the current files being parsed and checks if the file to be included isn't already
in the stack to avoid loop.

The array of target functions is now returned by getTargetFunctions() method to be inheritable and to ease unit testing.

Added unit tests for the weeFileConfig class for basic features, platform targets and file includes.