weeAutoload Cache
Speeds up things A LOT.
Currently a constant named WEE_AUTOLOAD_FILE is needed that defines
the filename to the autoload cache file. When it is defined, it will
never gets deleted by the framework. This means that if you add classes
they will be ignored. This is no problem in a production environment,
and in development the best solution is to make the file WEE_AUTOLOAD_FILE
read-protected so that it just gets ignored and include files are indexed
all the time.