Improve weeFrame event dispatch
The weeFrame event dispatch process has been rewritten: instead of a boolean-returning isAuthorized($aEvent) method, we now have
a setup($aEvent) method which can set things common to every event of the frame and fire an UnauthorizedAccessException if necessary.
When an UnauthorizedAccessException is thrown the unauthorizedAccess($aEvent) is called and can recover the unauthorized access by sending
a new event to weeApplication. If the unauthorized access is not recovered, weeApplication displays a page in the same way it does for a
missing wee conf file.