First version of the new and improved API viewer.
The viewer loads the XML file at startup, and then generate pages dynamically.
We think this is the best choice to make the API fast and reliable.
The viewer will also link to Trac's source code browser where applicable.

weeDocumentor has been much improved for this tool.
Also added the jquery history plugin to the resources.