Subsets support in weeDbSetScaffold.
Removed the search* methods in favor of subset support.
Subsets allow you to filter the rows in the table while still
using the object like you would use a normal set.

In addition, operations are supported between two sets A and B:
union, intersection, complement and symmetric difference,
allowing to dynamically perform set operations on the rows of the db.