Added addons manager; bump version to 0.6.0a.
The addons manager is still mostly a skeleton of what it should be,
but it works and it allows us to expand developments on many sides
by writing or including addons.

The version of the framework has been bumped to 0.6.0a in order to
test the addons manager.

The addons manager currently lacks a backend side where developers
will be able to register themselves and manage their projects. For
the time being the backend operations will be done by us manually.