Use weeRenderer in weeFrame
A frame now holds onto a renderer instead of a template, but
the mechanism to obtain a default `weeTemplate` instance based
on `sBaseTemplate`-related properties is still there, thus
keeping backward compatibility.

As a renderer may not always be an instance of `weeTemplate`, we
now check in the `set` method whether the current renderer associated
with the frame is also an instance of `weeDataHolder`.