Detect ambiguous fields in the WHERE and ORDER BY clauses.
The usefulness of this may vary depending on the RDBMS used,
but at least joins should work everywhere without problem now.
Some databases throw an error when a column is specified and
its name is found in 2 different tables of a join. This often
happens because foreign keys and primary keys share the same
name and data. By prefixing their name with the name of the
original table, we prevent this error from occuring.