Print a notice about required fields automatically.
Makes the UI of forms much better by printing a * in front of each required
fields. A note is added on top (but below the global errors if any) to
notify the user that some fields are required. To disable this behavior
the form class hide-required can be used.

In addition, ignore the fileinput widgets when validating the form. This
allows us to put a mark on those widgets while still having the validation
success.