Create method weeApplication::serveAsFile
This method permits to serve the output the frame of the application as
a file through header `Content-Disposition`. Remove the output method of
weeOutput and write its short code directly in weeApplication. This
snippet may be moved later to its very own function, but I feel
weeOutput need to be removed, the class is awkward with just two static
methods.