Remove weeTemplate::mkLink, use weeURL instead.
weeURL is a safer, context-aware, method to generate URLs.
It can be used from the template through the $this->url method,
or outside by creating a weeURL object that can then be passed
to the template. Outside the template the data isn't encoded
(unless it's passed an encoder), inside it is encoded correctly.
Data values from inside the template are considered encoded,
data from outside isn't.