weeDbSetScaffold: improve JOINs; allows custom JOINs and SELECTs.
A single criteria can now be used for more than 2 conditions, for example
"AND a b c d e" will mean "a AND b AND c AND d AND e".

By refactoring the buildJoin method, it is now possible to write more
adequate JOINs where it wasn't possible before. This should however be
used with great care.

SELECT expressions now are customizable by overriding buildSelectExpression.
It defaults to * like before, but can be used to add subqueries or handle
special JOIN cases like for example when you have 2 foreign keys pointing
to the same table.